for other plapforms %d == int
git-svn-id: http://svn.uvw.ru/mhddfs/trunk@88 d7d42532-9e51-4b1d-b199-9e4c94b77a30

+++ src/main.c
@@ -281,7 +281,7 @@ static int mhdd_internal_open(const char *file,
   if (fchown(fd, fcontext->uid, fcontext->gid)!=0)
   {
     mhdd_debug(MHDD_INFO, "mhdd_internal_open: error: can not set owner %d:%d to %s: %s\n",
-      fcontext->uid, fcontext->gid, path, strerror(errno));
+      (int)fcontext->uid, (int)fcontext->gid, path, strerror(errno));
   }
   struct files_info *add=add_file_list(file, path, fi->flags, fd);
   fi->fh=add->id;
